refactor: 메세지룸 아이템 분리
